+++ embed-test.html
@@ -10,7 +10,10 @@
 <body>
 
 
-    <div id="krets-container">
+    <div id="krets-thanks-container" class="krets-container">
+        Tusen takk 🎉
+    </div>
+    <div id="krets-main-container" class="krets-container">
         <ul id="krets-emoji-list">
             <ul class="flex-container">
                 <li class="krets-flex-element" role="button">
@@ -32,6 +35,7 @@
             <h2 id="krets-contact-header" class="krets-sub-header">Valgfri kontaktinformasjon</h2>
             <input id="krets-contact-input" class="krets-input" type="email" placeholder="e-post eller telefonnummer" />
             <button id="krets-send-button">send</button>
+            <p id="krets-error-message">auda, det skjedde en feil med tilbakemeldingen..</p>
         </div>
     </div>
 
@@ -41,10 +45,28 @@
             padding: 0;
         }
 
-        #krets-container {
+        .krets-container {
             display: flex;
             flex-direction: column;
             align-items: center;
+        }
+
+        #krets-thanks-container {
+
+            transition: 1s ease-in;
+            opacity: 0;
+            display: none;
+            text-align: center;
+            padding: 1em;
+            font-size: 5em;
+            text-decoration: underline;
+            text-decoration-color: hsl(183, 80%, 20%);
+
+        }
+
+        #krets-error-message {
+            display: none;
+            color: orangered;
         }
 
         .krets-sub-header {
@@ -184,6 +206,10 @@
         const getContactHeader = get("#krets-contact-header");
         const getSendButton = get("#krets-send-button");
 
+        const getThanksContainer = get("#krets-thanks-container")
+        const getMainContainer = get("#krets-main-container")
+        const getErrorMessage = get("#krets-error-message");
+
         const initializeState = (initializeState) => {
 
 
@@ -224,6 +250,17 @@
             getOptionalInuptContainer().style.opacity = 1;
         }
 
+        const showThanks = () => {
+
+            getMainContainer().style.display = "none";
+            getThanksContainer().style.opacity = 1;
+            getThanksContainer().style.display = "block";
+        }
+
+        const showError = () => {
+
+            getErrorMessage().style.display = "block";
+        }
 
         const onSend = async () => {
 
@@ -235,12 +272,6 @@
                 return;
             }
 
-            /*
-            token: string,
-            response: ResponseModel
-            answers: AnswerModel[]
-            */
-
             const embeddableResponse = {
                 token: TOKEN,
                 response: {
@@ -248,19 +279,21 @@
                     emotion,
                     contact_details: contactDetails,
                 },
-                answers: answers.values()
+                answers: Array.from(answers.values())
             }
 
-            const { status } = await fetch(`${API_BASE}/api/pages/${PAGE_ID}/embeddables`, {
-                method: "PUT",
+            console.log("Going to run fetch")
+            const { status } = await fetch(`${API_BASE}api/pages/${PAGE_ID}/embeddables/responses`, {
+                method: "POST",
                 headers: {
                     "Content-Type": "application/json"
                 },
                 body: JSON.stringify(embeddableResponse)
             });
 
-            console.log(status);
-
+            status === 201 ?
+                showThanks() :
+                showError();
         }
 
         const setupClickListeners = () => {
